# backend/ReelJett/ReelJett.WebAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ReelJett.WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Movie/GetNewReleaseMovies": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "moviesPerPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "language",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movie/GetUpcomingMovies": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "moviesPerPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "language",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movie/GetInterestedMovies": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieid",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "language",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movie/GetMovieDetails": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieid",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "title",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movie/GetSearchedMovie": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "type": "string"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "moviesPerPage",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movie/GetMovieEmbedLink": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieid",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "title",
            "type": "string"
          },
          {
            "in": "query",
            "name": "year",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movie/GetTrailer": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "movieid",
            "type": "string"
          },
          {
            "in": "query",
            "name": "language",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/PersonalMovie/AddPersonalMovie": {
      "post": {
        "tags": [
          "PersonalMovie"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddPersonalMovieDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/PersonalMovie/GetPersonalMovies": {
      "get": {
        "tags": [
          "PersonalMovie"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "moviesPerPage",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/PersonalMovie/DeletePersonalMovie": {
      "delete": {
        "tags": [
          "PersonalMovie"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AddPersonalMovieDTO": {
      "type": "object",
      "properties": {
        "link": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "poster": {
          "format": "byte",
          "type": "string"
        }
      }
    },
...
  }
}